delete flight when no route is present

+++ logbook/models.py
@@ -85,8 +85,9 @@ class Flight(models.Model):
         super(Flight,self).save(*args, **kwargs)
         
     def delete(self, *args, **kwargs):
-        from route.models import Route
-        Route.objects.get(flight__pk=self.pk).delete()
+        if self.route:
+            from route.models import Route
+            Route.objects.get(flight__pk=self.pk).delete()
         super(Flight,self).delete(*args, **kwargs)
         
     class Meta:
